+++ src/model/ScanNode.ts
@@ -18,12 +18,16 @@ export class ScanNode implements INode {
     public queries: any[] | undefined;
 
     constructor(public scanId: number, public projectId: number,
-        public readonly sourceLocation: string, public isFolder: boolean,
+        public readonly sourceLocation: string, public readonly isFolder: boolean,
         public readonly httpClient: HttpClient, private readonly log: Logger,
-        public parentNode: ServerNode) {
+        public parentNode: ServerNode, public readonly isScannedByVSC: boolean) {
         this.scanResult = new ScanResults();
-        if(parentNode.config){
+        if (parentNode.config) {
             this.scanResult.updateSastDefaultResults(parentNode.config.sastConfig);
+        } else {
+            this.scanResult.url = parentNode.sastUrl;
+            this.scanResult.enablePolicyViolations = false;
+            this.scanResult.thresholdEnabled = false;
         }
     }
 
@@ -93,13 +97,6 @@ export class ScanNode implements INode {
         return severityNodes;
     }
 
-    public canRetrieveResults(): boolean {
-        if (this.httpClient.accessToken) {
-            return true;
-        }
-        return false;
-    }
-
     public async retrieveScanResults() {
         try {
             await this.addStatisticsToScanResults();
@@ -111,7 +108,7 @@ export class ScanNode implements INode {
         }
     }
 
-    public async addStatisticsToScanResults() {
+    private async addStatisticsToScanResults() {
         const cxServer: CxServerSettings = CxSettings.getServer();
         const statistics: any = await this.httpClient.getRequest(`sast/scans/${this.scanId}/resultsStatistics`);
 
@@ -142,7 +139,7 @@ Scan results location:  ${this.scanResult.sastScanResultsLink}
 `);
     }
 
-    public async addDetailedReportToScanResults() {
+    private async addDetailedReportToScanResults() {
         const client = new ReportingClient(this.httpClient, this.log);
         if (!CxSettings.isQuiet()) {
             vscode.window.showInformationMessage('Waiting for server to generate scan report');
@@ -153,9 +150,6 @@ Scan results location:  ${this.scanResult.sastScanResultsLink}
         this.scanResult.scanTime = doc.$.ScanTime;
         this.scanResult.locScanned = doc.$.LinesOfCodeScanned;
         this.scanResult.filesScanned = doc.$.FilesScanned;
-        if(this.scanResult.filesScanned == 1){
-            this.isFolder = false;
-        }
         this.queries = doc.Query;
         this.scanResult.queryList = ScanNode.toJsonQueries(doc.Query);
         if (!CxSettings.isQuiet()) {
@@ -213,5 +207,4 @@ Scan results location:  ${this.scanResult.sastScanResultsLink}
             await CxSettings.updateReportPath(jsonReportPath);
         }
     }
-
 }