+++ CHANGELOG.md
@@ -1,10 +1,18 @@
-## 2020.2.1-BETA
+## 2020.3.1
 
-* First release as beta version
+* First release generally available
 * Bind an existing project for scan
 * Create a new project for scan
 * Execute CxSAST scan
-* Retrieve CxSAST scan results
+* Retrieve CxSAST results of scanned source code
 * Show vulnerabilities on Result Table and Attack Vector views
 * Save CxSAST scan report to external JSON file
 * Show vulnerability query description
+* Retrieve last scan results of bound project (without running a scan)
+* Ability to disable 'Scan Any File/Folder' buttons
+* Login - support credentials and SSO methods
+* Support incremental/full scans
+* Support Private/Public Scans
+* Unbind project
+* Silent mode - controls the amount of popup messages displayed to the user
+* Config as Code for selected attributes